requirements.txt
pandas==0.24.2
python-docx==0.8.10
